# medvision-cxr/frontend/public/openapi/medvision-cxr-openapi.yaml
openapi: 3.1.0
info:
  title: MedVision-CXR Backend
  version: 0.1.0
  summary: OpenAPI 3.1 specification for the MedVision-CXR backend
  description: MedVision-CXR is an AI-assisted chest X-ray triage system. This specification
    covers authentication, upload, AI-assisted risk assessment, Grad-CAM, doctor review,
    audit logs, model metadata, direct delete, health, and the governed deletion approval
    workflow.
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Get Me
      operationId: get_me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/audit/logs:
    get:
      tags:
      - audit
      summary: List audit logs
      description: 列出系统审计日志。支持按 action_type、resource_type、request_id、actor_user_id
        过滤。仅 doctor/admin 可访问。
      operationId: get_audit_logs_api_v1_audit_logs_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: action_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Type
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
      - name: request_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      - name: actor_user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Actor User Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Audit logs returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '403':
          description: Insufficient permissions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/audit/logs/{audit_log_id}:
    get:
      tags:
      - audit
      summary: Get audit log detail
      description: 获取单条审计日志详情。仅 doctor/admin 可访问。
      operationId: get_audit_log_detail_api_v1_audit_logs__audit_log_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: audit_log_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Audit Log Id
      responses:
        '200':
          description: Audit log returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogOut'
        '403':
          description: Insufficient permissions
        '404':
          description: Audit log not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuditLogListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AuditLogOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - items
      - total
      title: AuditLogListResponse
    AuditLogOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        actor_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Actor User Id
        action_type:
          type: string
          title: Action Type
        resource_type:
          type: string
          title: Resource Type
        resource_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Resource Id
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        ip_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Hash
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
        event_payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Event Payload
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - id
      - actor_user_id
      - action_type
      - resource_type
      - resource_id
      - request_id
      - ip_hash
      - user_agent
      - event_payload
      - created_at
      - updated_at
      title: AuditLogOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
...
    RoleOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        role_code:
          type: string
          title: Role Code
        role_name:
          type: string
          title: Role Name
      type: object
      required:
      - id
      - role_code
      - role_name
      title: RoleOut
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        disclaimer:
          type: string
          title: Disclaimer
          default: 本系统仅用于胸部 X 光影像的辅助筛查、辅助分诊和医生复核优先级排序，不用于自动诊断，不替代医生，不提供治疗建议。
      type: object
      required:
      - access_token
      - expires_in
      title: Token
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        role_code:
          type: string
          title: Role Code
          default: clinician
      type: object
      required:
      - email
      - password
      title: UserCreate
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        status:
          type: string
          title: Status
        role:
          $ref: '#/components/schemas/RoleOut'
      type: object
      required:
      - id
      - email
      - status
      - role
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
servers:
- url: http://localhost:8000
  description: Local development server